Create Handbook System and set page size for get api to 50
