+++ Holoreact/Assets/Script/HandBookManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39d0d5f0015335c4a9337fd394a1523e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
